Find links in SugarCube-style <<click>> and <<link>> macros
The <<link>> syntax is ambiguous in the 2-argument case. I checked the
SugarCube source and if the last argument is a known action keyword,
it is treated as an action, otherwise it is treated as a location.

Since these macros are SugarCube-specific, this code should probably
be moved into the header-specific Python module, once that lands.